+++ projects/modular_llm/finetune_experts.py
@@ -9,7 +9,7 @@ sys.path.append(os.path.join(os.path.dirname(__file__), "..", ".."))
 
 from typing import Callable
 
-from mttl.callbacks import LiveCheckpointCallback, RougeCallback
+from mttl.callbacks import DownstreamEvalCallback, LiveCheckpointCallback, RougeCallback
 from mttl.config import FinetuneConfig
 from mttl.datamodule.base import get_datamodule
 from mttl.logging import get_pl_loggers, logger, setup_logging
@@ -33,7 +33,6 @@ from mttl.models.library.library_transforms import (
 from mttl.models.modifiers.base import ModifierConfig
 from mttl.models.monitors import get_monitors
 from mttl.utils import get_checkpoint_path, remote_login
-from projects.modular_llm.src.callbacks import DownstreamEvalCallback
 from projects.modular_llm.src.retrievers import RandomRetriever, SVDEmbeddingRetriever
 
 FINETUNE_FUNCTIONS: dict[str, Callable] = {}
