+++ projects/modular_llm/train_moe_main.py
@@ -6,14 +6,18 @@ from pytorch_lightning import Trainer, seed_everything
 
 sys.path.append(os.path.join(os.path.dirname(__file__), "..", ".."))
 
-from mttl.callbacks import LiveCheckpointCallback, NanoMMLUCallback, RougeCallback
+from mttl.callbacks import (
+    DownstreamEvalCallback,
+    LiveCheckpointCallback,
+    NanoMMLUCallback,
+    RougeCallback,
+)
 from mttl.config import ExpertConfig
 from mttl.datamodule.base import get_datamodule
 from mttl.logging import get_pl_loggers, logger, setup_logging
 from mttl.models.expert_model import MoEModel
 from mttl.models.monitors import get_monitors
 from mttl.utils import remote_login
-from projects.modular_llm.src.callbacks import DownstreamEvalCallback
 
 
 def run_multitask(args: ExpertConfig):
